Refactor aws components
